chore: separate artefact and image services
Image service uses ArtefactService to pull build files, but is responsible for pushing any images
