+++ internal/download/download.go
@@ -73,7 +73,7 @@ func Artefacts(ctx context.Context, client adminpbconnect.AdminServiceClient, ke
 }
 
 // ArtefactsFromOCI downloads artefacts for a deployment from an OCI registry.
-func ArtefactsFromOCI(ctx context.Context, client ftlv1connect.SchemaServiceClient, key key.Deployment, dest string, service *oci.OCIArtefactService) error {
+func ArtefactsFromOCI(ctx context.Context, client ftlv1connect.SchemaServiceClient, key key.Deployment, dest string, service *oci.ArtefactService) error {
 	logger := log.FromContext(ctx)
 	response, err := client.GetDeployment(ctx, connect.NewRequest(&ftlv1.GetDeploymentRequest{
 		DeploymentKey: key.String(),
